feat: encapsulate container bindings merging within framework
