+++ packages/core/framework/src/config/index.ts
@@ -1,3 +1,5 @@
+import "./types/container"
+
 export * from "./loader"
 export * from "./config"
 export * from "./types"
