+++ packages/core/framework/src/database/index.ts
@@ -1,1 +1,2 @@
+import "./types/container"
 export * from "./pg-connection-loader"
