+++ packages/core/framework/src/feature-flags/index.ts
@@ -1,3 +1,5 @@
+import "./types/container"
+
 export * from "./types"
 export * from "./feature-flag-loader"
 export * from "./flag-router"
