+++ packages/core/framework/src/utils/index.ts
@@ -1,1 +1,3 @@
+import "./types/container"
+
 export * from "@medusajs/utils"
