+++ packages/core/framework/src/workflows-sdk/index.ts
@@ -1,1 +1,3 @@
+import "./types/container"
+
 export * from "@medusajs/workflows-sdk"
