+++ packages/core/framework/src/workflows/index.ts
@@ -1,1 +1,3 @@
+import "./types/container"
+
 export * from "./workflow-loader"
